requirements.txt
fastapi
pydantic
requests
uvicorn
pandas
openpyxl
streamlit
groq
python-dotenv
